% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champagne_glass_for.R
\name{champagne_glass_for}
\alias{champagne_glass_for}
\title{Vectorized Champagne Glass Calculation using a Loop}
\usage{
champagne_glass_for(x)
}
\arguments{
\item{x}{A numeric value or vector of heights at which to calculate the glass radius.}
}
\value{
A numeric vector of radii corresponding to the heights given by `x`.
}
\description{
This function calculates the side profile of a champagne glass using a for loop for vectorized input.
}
\examples{
champagne_glass_for(seq(0, 20, by = 1))
}
