% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champagne_glass_vectorized.R
\name{champagne_glass_vectorized}
\alias{champagne_glass_vectorized}
\title{Vectorized Calculation of Champagne Glass Profile}
\usage{
champagne_glass_vectorized(x)
}
\arguments{
\item{x}{A numeric vector of heights at which to calculate the glass radius.}
}
\value{
A numeric vector of radii corresponding to the heights given by `x`.
}
\description{
This function provides a vectorized version of the \code{champagne_glass} function
to calculate the profile of a champagne glass over a vector of heights.
}
\examples{
champagne_glass_vectorized(c(-1, 0.25, 5, 12, 16, 21))
}
